requirements.txt
# python==3.7.6
tensorflow-gpu==1.15.3
scipy==1.5.2
tqdm==4.48.2